check_packages : improve the "skip klibc dep" hack.
It is not possible to check the klibc dep on the PKGBUILD level, so I made a
hack to skip it.
This hack broke on klibc-jfflyAahxqaliwAofrf_fdf5upI because of the
underscore.
But we can simply use the \w regexp which matches any alpha-numeric char,
including underscore. According to klibc developer, dashes are possible too.
Also the length of this string is always 27 chars, so we can use that.

Signed-off-by: Xavier Chantry <shiningxc@gmail.com>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>