Remove chmod on /home/aur/*
This one portion of the adjust-permissions script takes over a minute of
wall clock time to run on gerolde when no other cron jobs are running. It
takes three minutes when we hit the hour mark on the clock due to everything
else going at the same time. A find on /home/aur/unsupported reveals this
command will need to touch ~77000 files.

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>