check_packages: fix stupid assumption on script location.
The parse_pkgbuilds.sh script was assumed to be in the current working
directory, which is quite stupid since check_packages.py can be called from
anywhere. Now it only assumes that check_packages.py and parse_pkgbuilds.sh
are in the same directory.

This should fix the empty integrity checks on arch-dev :)

Signed-off-by: Xavier Chantry <shiningxc@gmail.com>
Signed-off-by: Dan McGee <dan@archlinux.org>