Be more careful with config file inclusion
We can't really do the basedir magic from db-functions as it is just being
sourced and is not the currently executing script. Although a bit
repetitive, it is a lot safer to just include the config file everywhere.

Noticed this when trying to run the latest available scripts on sigurd.

Signed-off-by: Dan McGee <dan@archlinux.org>