check_packages : fix a bug when filling provisions
When we had a duplicate package, we would replace the old ref by the new
ref. But the provisions dict kept both the old and new ref. This was not
good at all.

Now, we just keep the old ref, and we only fill provisions after all
packages have been parsed. This should be much more sane.

Signed-off-by: Xavier Chantry <shiningxc@gmail.com>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>