Added new cronjob script to clean up unneeded sourceballs
This script is intended to be a weekly cron-job.  It supplements the current sourceballs cleanup in make-sourceball that only removes the old sourceballs.  This script removes the sourceballs of packages that were either removed completely from the repo or moved to a splitted package. It also checks the license of the packages which have a sourceball to see, in case of a license change, if the hosting of the sources is still necessary.

Signed-off-by: Eric Bélanger <snowmaniscool@gmail.com>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>