create-filelists: rework the package loop completely
Instead of wasting time extracting .PKGINFO twice from every single package
in the repos, use the package DB to eliminate most of the heavy lifting.
This way we only need to worry about looking at the packages that actually
have changed since the last time we built the package database.

This should give a noticeable performance increase to this job in addition to
reducing IO load and unnecessary reading of every package file.

Signed-off-by: Dan McGee <dan@archlinux.org>