Make repo locking an atomic process and added timeout argument
The repo locking function now use a lock directory instead of a lock
file.  This makes the lock checking and creation an atomic process.
The repo_lock function will now try to obtain a lock every $LOCK_DELAY
seconds until it is successful.  An optional third argument can be use
to give a timeout in seconds; this is intended for scripts that runs
unattended.  Repo locking is now used in the ftpdir-cleanup script.
This should fix the problem of the ftpdir-cleanup script removing the
new package instead of the old one (FS#17058).

Signed-off-by: Eric Bélanger <snowmaniscool@gmail.com>