Add update-web-db cronjob
This has been untracked by version control for a while on gerolde in both
the cron-jobs directory and in /etc/cron.hourly/. Add it here so we can make
changes and know what is going on.

This is an improved script over what we currently have. It is one script
instead of two, and it does things a little smarter with the logging.

Signed-off-by: Dan McGee <dan@archlinux.org>