db-*: always handle pkgname/pkgbase/pkgfile argument last
This matches what we do in db-update, and makes the most sense if we
want to expand these commands to take multiple arguments at the same
time. Especially in the case of db-repo-add and db-repo-remove, these
wrapper commands should be no more limiting than repo-add and
repo-remove which always allow multiple arguments and are more efficient
that way.

This patch simply reorders the arguments; a later patch will support
multiple pkgname/pkgbase/pkgfile arguments on the command line.

Although this does break backwards compatibility and requires some
changing of muscle memory, I think it is a worthwhile change to make as
the functionality of these commands becomes more powerful and the order
of arguments becomes the same in all of them.

Signed-off-by: Dan McGee <dan@archlinux.org>