test: Don't use "! cmd" except as the last statement in a function
With BATS setting up error traps, we get used to writing simple `[[ foo ]]`
assertions and not having to check the result.  However, using `!` to
invert the result DOESN'T trigger the trap.  It works OK when it is the
last command in a function, as it still affects the function's return
value, so then the trap triggers in the caller (rather than in the
function).  This means that a check may run successfully when it should
fail.

So, replace all uses of bare `! cmd` with `if cmd; then return 1; fi`,
unless it is the final statement in a function (as it is in
sourceballs.bats:__checkRemovedSourcePackage())

The mistake meant that checkRemovedPackage() was effectively equivalent to
checkRemovedPackageDB().  This meant that no one noticed that db-updates's
"add package with inconsistent name fails" test called checkRemovedPackage
instead of checkRemovedPackageDB.  Fixing the ! issue means that the test
now fails, so change which function it calls.