db-update: performance optimization when checking if a buildinfo exists
Use the --fast-read option to quit as soon as the file is first found. A
package file should never contain two of these anyway, but even if it
did, we simply want to know that one exists!

makepkg guarantees its internal .METADATA files are tar'ed up first, so
this should usually be quite rapidly found. For very large packages,
crawling the entire package file to check for a later entry overriding
the first one, is... pointless... and worst of all, slow.

This reflects the identical mode used for extracting the pkginfo file,
which was inconsistently implemented from the initial feature addition.