test/lib/common: Make checkRemovedPackage aware of debug packages
This ensures we do not have the debug repository at hand when looking at
the SVN state files. However, there is no good way to figure out if we
have a debug package as .BUILDINFO nor .PKGINFO has the information we
need.

We blindly assume that any package where we look up a *-debug repository
is a debug package.

Signed-off-by: Morten Linderud <morten@linderud.pw>