cronjobs/ftpdir-cleanup: Support DEBUGPKGPOOL
This implementation is probably the easiest one and/or cheap. After
building a list of all packages in the repositories we filter the lists
into two separate working lists. One for normal packages and one for
debug packages.

The only issue is that we can encounter globs with no matching cleanup
lists, like when running ftpdir-cleanup after the debug package changes
has been introduced, but before deploying new repositories. We guard
against this by ensuring we have found debug packages, and normal
packages, before checking the working lists.

This introduces some duplicated code but should be fine. Rest of the
logic is mirrored between the two package pools.

Signed-off-by: Morten Linderud <morten@linderud.pw>