test: fix release helper by avoiding duplicate tags on the same commit
The way a lot of test cases are build collides with the way our git
packaging works. We tag every release in the API, which will fail
for some of those tests setups as the tag already exists.
Fix this issue by only trying to tag a version if the revision we try to
tag actually differs from an existing tag. This allows to preserve the
easy going test cases as the functional ideology of the test case won't
be compromised by this test check.