db-update: fix retrieving metadata from any-packages for multi arch
For any-packages on a multi arch repo setup, we can't call `getpkgbase`
or `getpkgver` on the original `$pkg` location, as they get moved to a
different location during per arch repo target handling. Instead of
trying to refactor all the behavior of doing so, lets instead choose to
retrieve this information once as first action and only access the
cached metadata afterwards.

Before this commit, the cache has only been populated as last action of
the update process, which will fail horribly for the second set of
packages targeting another architecture's repo.