implemented unix group based ACL to shield packaging repositories
Introduce an associate array that can be set in the config to define the
permissions per group. Declare permission sub sets so we can easier
puzzle together the actual permissions per group without duplicating
all of the repos in the declaration.

Furthermore we specifically use `packager` and `junior-packager` instead
of `tu` to reflect the rename on system group level.

Reflect second batch of changes related to RFC-14 (Merge package
repositories).

https://gitlab.archlinux.org/archlinux/rfcs/-/merge_requests/14