git: purely operate on a cached bare mirror instead of checkouts
Even with depth=1 we are better of to simply keep caches of the clones.
Introduce a new `config` location to store all repositories inside.
Whenever the checkout function is called, we now either clone the
repository or fetch the latest deltas.

Functions that require the actual PKGBUILD are not operating purely in
memory and with virtual file handles instead of dealing with actual file
checkouts.