db-update: always pass both, signature and file to gpg verification
Only specifying the signature and having gpg rely on assumptions about
what to check the signature against is a very bad practice leading to
ambiguous possibilities if the signature contains data itself. Make
sure that we do not call gpg verification like such by passing both, the
signature as well as the actual file of the data to check against.