db-update: fix state repo entries for any packages
We do not serve any packages as dedicated architecture repository in
pacman, but that doesn't mean we must change the current behavior of
dbscripts. A lot of places are build in a way to expect any packages
to actually have a dedicated any entry of the matching repository.
Lets not try to change this behavior while migrating to git as we
already face too much possibility for issues. Hence map any packages
to any when changing the state repository.