git: simplify state handling by not requiring paths to pre exist
It's a lot more convenient for operation if we do not require pre
existing state assumptions but rather ensure we initialize whatever
path we require during execution of the vcs state mutation functions.
Any information we want to retrieve from the history will also not
require any paths to exist indefinitely, but instead we follow paths
in the history no matter when they appear or disappear in between.