db-move: fix repo mutation when testing packages overwrite old stables
We can safely call with `--force` to force renaming or moving of a file
even if the target exists. This is a common case whenever we want to
move packages between repositories which will result in overwriting an
old stable version from testing.
This commit makes a bunch of new tests cases pass and also makes
logically a lot of sense to do.