test: collapse container layers that do not need to be individual
It is discouraged to have every command in an own container layer if
there is no benefit in having them separate. For example it makes sense
to have a static block before having a dynamic block like system updates
hence the first layer will mostly never be rebuild. But that all isn't
much required here, lets keep it simple by collapsing the polluted
amount of layers into the minimum set.